Fix heroku issues
