+++ server/config/config.js
@@ -1,16 +1,16 @@
 import firebase from 'firebase';
 import dotenv from 'dotenv';
 
-dotenv.config();
+// dotenv.config();
 
-const config = {
-  apiKey: process.env.apiKey,
-  authDomain: process.env.authDomain,
-  databaseURL: process.env.databaseURL,
-  projectId: process.env.projectId,
-  storageBucket: process.env.storageBucket,
-  messagingSenderId: process.env.messagingSenderId
-};
-const dbConfig = firebase.initializeApp(config);
+// const config = {
+//   apiKey: process.env.apiKey,
+//   authDomain: process.env.authDomain,
+//   databaseURL: process.env.databaseURL,
+//   projectId: process.env.projectId,
+//   storageBucket: process.env.storageBucket,
+//   messagingSenderId: process.env.messagingSenderId
+// };
+// const dbConfig = firebase.initializeApp(config);
 
-export default dbConfig;
+// export default dbConfig;
