+++ server/middleware/authentication.js
@@ -1,12 +1,22 @@
 import dbConfig from '../config/config';
 
-const authenticateUser = () => new Promise((resolve) => {
-  dbConfig.auth().onAuthStateChanged((currentUser) => {
-    if (currentUser) {
-      resolve(currentUser);
-    }
-    resolve({});
-  });
-});
+// const authenticateUser = () => new Promise((resolve) => {
+//   dbConfig.auth().onAuthStateChanged((currentUser) => {
+//     if (currentUser) {
+//       resolve(currentUser);
+//     }
+//     resolve({});
+//   });
+// });
 
-export default authenticateUser;
+const isAuthenticated = (req, res, next) => {
+  const user = dbConfig.auth().currentUser;
+  if (user !== null) {
+    req.user = user;
+    next();
+  } else {
+    res.redirect('/signIn');
+  }
+};
+
+export default isAuthenticated;
