+++ test/mocha.conf
@@ -1,0 +1,8 @@
+--colors
+--reporter spec
+--timeout 1000
+--check-leaks
+--compilers js:babel-core/register
+--require babel-polyfill
+--recursive
+--sort
